Detect virtual env in opencv/install_deb.sh

+++ packages/cv/opencv/install_deb.sh
@@ -53,23 +53,27 @@ bash /tmp/cmake/install.sh
 bash /tmp/numpy/install.sh
 
 # manage some install paths
-PYTHON3_VERSION=`python3 -c 'import sys; version=sys.version_info[:3]; print("{0}.{1}".format(*version))'`
-
 if [ $ARCH = "aarch64" ]; then
 	local_include_path="/usr/local/include/opencv4"
-	local_python_path="/usr/local/lib/python${PYTHON3_VERSION}/dist-packages/cv2"
 
 	if [ -d "$local_include_path" ]; then
 		echo "$local_include_path already exists, replacing..."
 		rm -rf $local_include_path
 	fi
 
+	ln -sfnv /usr/include/opencv4 $local_include_path
+fi
+
+VIRTUAL_ENV=`python3 -c "import sys; print('1' if sys.prefix != sys.base_prefix else '')"`
+
+if [[ ! -n "$VIRTUAL_ENV" ]] && [[ $ARCH = "aarch64" ]]; then
+	PYTHON3_VERSION=`python3 -c 'import sys; version=sys.version_info[:3]; print("{0}.{1}".format(*version))'`
+	local_python_path="/usr/local/lib/python${PYTHON3_VERSION}/dist-packages/cv2"
+
 	if [ -d "$local_python_path" ]; then
 		echo "$local_python_path already exists, replacing..."
 		rm -rf $local_python_path
 	fi
 
-	ln -sfnv /usr/include/opencv4 $local_include_path
 	ln -sfnv /usr/lib/python${PYTHON3_VERSION}/dist-packages/cv2 $local_python_path
 fi
-
